refactor: reduces precision of overall layout cached values

+++ source/Age/Elements/RectEdges.cs
@@ -2,13 +2,13 @@ namespace Age.Elements;
 
 public record struct RectEdges
 {
-    public uint Top;
-    public uint Right;
-    public uint Bottom;
-    public uint Left;
+    public ushort Top;
+    public ushort Right;
+    public ushort Bottom;
+    public ushort Left;
 
-    public readonly uint Horizontal => this.Left + this.Right;
-    public readonly uint Vertical   => this.Top + this.Bottom;
+    public readonly uint Horizontal => (uint)(this.Left + this.Right);
+    public readonly uint Vertical   => (uint)(this.Top + this.Bottom);
 
     public override readonly string ToString() =>
         $"{{ Top: {this.Top}, Right: {this.Right}, Bottom: {this.Bottom}, Left: {this.Left} }}";
